Changed packet to short and moved variables

+++ AndroidApp/app/src/main/java/khr/easv/pokebotbroadcaster/app/data/BluetoothConnector.java
@@ -36,7 +36,7 @@ public class BluetoothConnector {
     }
 
     public void sendCommand( short command ) throws IOException {
-        _output.writeInt(command);
+        _output.writeShort(command);
         _output.flush();
     }
